+++ packages/apps/dev-wallet/src/App/app.tsx
@@ -1,21 +1,27 @@
+import { MediaContextProvider } from '@/Components/Media/media';
 import { DatabaseProvider } from '@/modules/db/db.provider';
 import { LayoutProvider } from '@/modules/layout/layout.provider.tsx';
 import { WalletProvider } from '@/modules/wallet/wallet.provider';
-import { useTheme } from '@kadena/kode-ui';
+import { useEffect } from 'react';
 import { PromptProvider } from '../Components/PromptProvider/Prompt';
 import { BetaHeader } from './BetaHeader';
 import { Routes } from './routes';
 import { SessionProvider } from './session';
 
 function Providers({ children }: { children: React.ReactNode }) {
-  // initialize the theme
-  useTheme({ overwriteTheme: 'dark' });
+  useEffect(() => {
+    if (!localStorage.getItem('theme')) {
+      localStorage.setItem('theme', 'dark');
+    }
+  }, []);
   return (
     <SessionProvider>
       <DatabaseProvider>
         <PromptProvider>
           <WalletProvider>
-            <LayoutProvider>{children}</LayoutProvider>
+            <MediaContextProvider>
+              <LayoutProvider>{children}</LayoutProvider>
+            </MediaContextProvider>
           </WalletProvider>
         </PromptProvider>
       </DatabaseProvider>