+++ packages/apps/dev-wallet/src/Components/CompactTable/utils/formatAmount.tsx
@@ -1,0 +1,10 @@
+import type { FC } from 'react';
+
+interface IProps {
+  value: string;
+}
+
+export const FormatAmount = (): FC<IProps> => {
+  const Component: FC<IProps> = ({ value }) => <>{value} KDA</>;
+  return Component;
+};