+++ packages/apps/dev-wallet/src/modules/network/network.repository.ts
@@ -1,9 +1,12 @@
+import { execInSequence } from '@/utils/helpers';
 import { IDBService, dbService } from '../db/db.service';
 
 export interface INetwork {
   uuid: string;
   networkId: string;
   name?: string;
+  default?: boolean;
+  disabled?: boolean;
   hosts: Array<{
     url: string;
     submit: boolean;
@@ -13,7 +16,8 @@ export interface INetwork {
 }
 
 export interface NetworkRepository {
-  getNetworkList: () => Promise<INetwork[]>;
+  getEnabledNetworkList: () => Promise<INetwork[]>;
+  getAllNetworks: () => Promise<INetwork[]>;
   getNetwork: (networkId: string) => Promise<INetwork>;
   addNetwork: (network: INetwork) => Promise<void>;
   updateNetwork: (network: INetwork) => Promise<void>;
@@ -28,9 +32,15 @@ const createNetworkRepository = ({
   remove,
 }: IDBService): NetworkRepository => {
   return {
-    getNetworkList: async (): Promise<INetwork[]> => {
+    getEnabledNetworkList: async (): Promise<INetwork[]> => {
+      const list: INetwork[] = await getAll('network');
+      return list.filter((network) => !network.disabled);
+    },
+
+    getAllNetworks: async (): Promise<INetwork[]> => {
       return getAll('network');
     },
+
     getNetwork: async (uuid: string): Promise<INetwork> => {
       return getOne('network', uuid);
     },
@@ -50,14 +60,16 @@ const createNetworkRepository = ({
 };
 export const networkRepository = createNetworkRepository(dbService);
 
-export const addDefaultNetworks = async () => {
-  const networks = await networkRepository.getNetworkList();
+export const addDefaultNetworks = execInSequence(async () => {
+  const networks = await networkRepository.getAllNetworks();
 
   if (!networks.find((network) => network.networkId === 'mainnet01')) {
     await networkRepository.addNetwork({
       uuid: crypto.randomUUID(),
       networkId: 'mainnet01',
       name: 'Mainnet',
+      // make mainnet disabled for now
+      disabled: true,
       hosts: [
         {
           url: 'https://api.chainweb.com',
@@ -73,6 +85,7 @@ export const addDefaultNetworks = async () => {
       uuid: crypto.randomUUID(),
       networkId: 'testnet04',
       name: 'Testnet',
+      default: true,
       hosts: [
         {
           url: 'https://api.testnet.chainweb.com',
@@ -87,7 +100,7 @@ export const addDefaultNetworks = async () => {
     await networkRepository.addNetwork({
       uuid: crypto.randomUUID(),
       networkId: 'testnet05',
-      name: 'Testnet-Next',
+      name: 'Testnet(Pact5)',
       hosts: [
         {
           url: 'http://api1.testnet05.chainweb.com',
@@ -98,4 +111,4 @@ export const addDefaultNetworks = async () => {
       ],
     });
   }
-};
+});