+++ packages/apps/dev-wallet/src/modules/wallet/wallet.hook.tsx
@@ -8,6 +8,7 @@ import * as AccountService from '../account/account.service';
 import { BIP44Service } from '../key-source/hd-wallet/BIP44';
 import { ChainweaverService } from '../key-source/hd-wallet/chainweaver';
 import { keySourceManager } from '../key-source/key-source-manager';
+import { INetwork } from '../network/network.repository';
 import { ExtWalletContextType, WalletContext } from './wallet.provider';
 import { IKeySource, IProfile } from './wallet.repository';
 import * as WalletService from './wallet.service';
@@ -243,8 +244,10 @@ export const useWallet = () => {
     askForPassword,
     getPublicKeyData,
     unlockKeySource,
-    setActiveNetwork,
+    setActiveNetwork: (network: INetwork) =>
+      setActiveNetwork ? setActiveNetwork(network) : undefined,
     activeNetwork: context.activeNetwork,
+    networks: context.networks,
     isUnlocked: isUnlocked(context),
     profile: context.profile,
     profileList: context.profileList ?? [],
@@ -252,6 +255,6 @@ export const useWallet = () => {
     keysets: context.keysets || [],
     keySources: context.keySources || [],
     fungibles: context.fungibles || [],
-    syncAllAccounts,
+    syncAllAccounts: () => (syncAllAccounts ? syncAllAccounts() : undefined),
   };
 };